pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "regime-detection"
version = "0.1.0"
description = "Multi-Modal Market Regime Analysis Framework v3.1"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Steven" },
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "hmmlearn>=0.3",
    "ruptures>=1.1",
    "scipy>=1.11",
]

[project.optional-dependencies]
fathon = [
    "fathon>=1.3",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/regime_detection"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
